AI manager subsystem
The subsystem contains an array of pointers to all AI actors. AIs can use this to communicate to update the current last seen position of the player. AIs register with this subsystem at runtime and unregister on death. Much better than each AI finding all other AIs at the start of the game and uses less memory.